feature/targeting_refactor
